WIP removing modules
