--- ios/Modules/Features/AddMeal/README.md
@@ -1,3 +1,0 @@
-# AddMeal
-
-A description of this package.
