--- ios/Modules/Features/AddMeal/Tests/AddMealTests/AddMealTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import AddMeal
-
-final class AddMealTests: XCTestCase {
-    func testExample() {
-        XCTAssertTrue(true)
-    }
-}
