--- ios/Modules/Features/Login/README.md
@@ -1,3 +1,0 @@
-# Login
-
-A description of this package.
