--- ios/Modules/Features/Login/Tests/LoginTests/LoginTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import Login
-
-final class LoginTests: XCTestCase {
-    func testExample() {
-        XCTAssertTrue(true)
-    }
-}
