--- ios/Modules/Features/MealList/Package.swift
@@ -1,24 +1,0 @@
-// swift-tools-version:5.3
-
-import PackageDescription
-
-let package = Package(
-    name: "MealList",
-    platforms: [
-        .iOS(.v14)
-    ],
-    products: [
-        .library(name: "MealList", targets: ["MealList"])
-    ],
-    dependencies: [
-        .package(path: "../Shared/Extensions"),
-        .package(path: "../Shared/Location"),
-        .package(path: "../Shared/Components"),
-        .package(path: "../Shared/AddMeal"),
-        .package(name: "shared", path: "../../../../shared/swiftpackage")
-    ],
-    targets: [
-        .target(name: "MealList", dependencies: ["shared", "Extensions", "Location", "Components", "AddMeal"]),
-        .testTarget(name: "MealListTests", dependencies: ["MealList"])
-    ]
-)
