--- ios/Modules/Features/MealList/README.md
@@ -1,3 +1,0 @@
-# MealList
-
-A description of this package.
