--- ios/Modules/Features/MealList/Sources/MealList/MapView.swift
@@ -1,64 +1,0 @@
-//
-//  MapView.swift
-//  iosHackthonApp
-//
-//  Created by Gareth Miller on 12/11/2020.
-//  Copyright © 2020 orgName. All rights reserved.
-//
-
-import SwiftUI
-import MapKit
-
-struct MapView: UIViewRepresentable {
-
-    var annotations: [MKPointAnnotation]
-    @Binding var selectedPlace: MKPointAnnotation?
-
-    func makeUIView(context: Context) -> MKMapView {
-        let mapView = MKMapView()
-        mapView.delegate = context.coordinator
-        mapView.showsUserLocation = true
-        mapView.region = MKCoordinateRegion(center: mapView.userLocation.coordinate, latitudinalMeters: 1000, longitudinalMeters: 1000)
-        mapView.userTrackingMode = .follow
-        mapView.isZoomEnabled = true
-        mapView.isPitchEnabled = true
-        mapView.isScrollEnabled = true
-        return mapView
-    }
-
-    func updateUIView(_ view: MKMapView, context: Context) {
-        if annotations.count != view.annotations.count {
-            view.removeAnnotations(view.annotations)
-            view.addAnnotations(annotations)
-        }
-    }
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(self)
-    }
-
-    class Coordinator: NSObject, MKMapViewDelegate {
-        var parent: MapView
-        var centerLocationOnce = true
-
-        init(_ parent: MapView) {
-            self.parent = parent
-        }
-
-        func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
-            guard !annotation.isKind(of: MKUserLocation.self) else {
-                return nil
-            }
-            let identifier = "meal"
-            var annotationView = mapView.dequeueReusableAnnotationView(withIdentifier: identifier)
-
-            if annotationView == nil {
-                annotationView = MKPinAnnotationView(annotation: annotation, reuseIdentifier: identifier)
-                annotationView?.canShowCallout = true
-            } else {
-                annotationView?.annotation = annotation
-            }
-            return annotationView
-        }
-    }
-}
