--- ios/Modules/Features/MealList/Tests/MealListTests/MealListTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import MealList
-
-final class MealListTests: XCTestCase {
-    func testExample() {
-        XCTAssertTrue(true)
-    }
-}
