--- ios/Modules/Features/Settings/README.md
@@ -1,3 +1,0 @@
-# Settings
-
-A description of this package.
