--- ios/Modules/Features/Settings/Sources/Settings/SettingsView.swift
@@ -1,23 +1,0 @@
-//
-//  SettingsView.swift
-//  
-//
-//  Created by Gareth Miller on 24/11/2020.
-//
-
-import SwiftUI
-
-public struct SettingsView: View {
-  
-  public init() { }
-  
-  public var body: some View {
-    Text("Settings screen")
-  }
-}
-
-struct SwiftUIView_Previews: PreviewProvider {
-  static var previews: some View {
-    SettingsView()
-  }
-}
