--- ios/Modules/Features/Settings/Tests/SettingsTests/SettingsTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import Settings
-
-final class SettingsTests: XCTestCase {
-    func testExample() {
-        XCTAssertTrue(true)
-    }
-}
