--- ios/Modules/Features/SignUp/README.md
@@ -1,3 +1,0 @@
-# SignUp
-
-A description of this package.
