--- ios/Modules/Features/SignUp/Tests/SignUpTests/SignUpTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import SignUp
-
-final class SignUpTests: XCTestCase {
-    func testExample() {
-        XCTAssertTrue(true)
-    }
-}
