--- ios/Modules/Features/TabBar/README.md
@@ -1,3 +1,0 @@
-# TabBar
-
-A description of this package.
