--- ios/Modules/Features/TabBar/Tests/TabBarTests/TabBarTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import TabBar
-
-final class TabBarTests: XCTestCase {
-    func testExample() {
-        XCTAssertTrue(true)
-    }
-}
