+++ ios/iosHackthonApp/Screens/Settings/SettingsView.swift
@@ -1,0 +1,23 @@
+//
+//  SettingsView.swift
+//  
+//
+//  Created by Gareth Miller on 24/11/2020.
+//
+
+import SwiftUI
+
+struct SettingsView: View {
+  
+  init() { }
+  
+  var body: some View {
+    Text("Settings screen")
+  }
+}
+
+struct SwiftUIView_Previews: PreviewProvider {
+  static var previews: some View {
+    SettingsView()
+  }
+}
