+++ shared/build.gradle.kts
@@ -5,7 +5,6 @@ plugins {
     kotlin("plugin.serialization")
     id("com.android.library")
     id("kotlin-android-extensions")
-    id("com.chromaticnoise.multiplatform-swiftpackage") version "1.0.2"
 }
 group = "com.kcc.kmmhackathon"
 version = "1.0-SNAPSHOT"
@@ -76,12 +75,6 @@ android {
         getByName("release") {
             isMinifyEnabled = false
         }
-    }
-}
-multiplatformSwiftPackage {
-    swiftToolsVersion("5.3")
-    targetPlatforms {
-        iOS { v("13") }
     }
 }
 val packForXcode by tasks.creating(Sync::class) {
